added EventManager class to allow global events for changes to trigger recalculation

+++ Assets/Scripts/Waypoint.cs
@@ -6,13 +6,13 @@ public class Waypoint
 {
     public Vector3 position;
     public Vector3 rotation;
-    public float time;
+    public int time;
     public Vector3 drag;
     public Vector3 lift;
     public Vector3 gravity;
     public Vector3 thrust;
 
-    public Waypoint(Vector3 position, Vector3 rotation, Vector3 drag, Vector3 lift, Vector3 gravity, Vector3 thrust, float time) {
+    public Waypoint(Vector3 position, Vector3 rotation, Vector3 drag, Vector3 lift, Vector3 gravity, Vector3 thrust, int time) {
         this.position = position;
         this.rotation = rotation;
         this.drag = drag;
@@ -22,7 +22,7 @@ public class Waypoint
         this.time = time;
     }
 
-    public Waypoint(Vector3 position, Vector3 rotation, float time): this(position, rotation, new Vector3(0f, 0f, 0f), new Vector3(0f, 0f, 0f), new Vector3(0f, 0f, 0f), new Vector3(0f, 0f, 0f), time)
+    public Waypoint(Vector3 position, Vector3 rotation, int time): this(position, rotation, new Vector3(0f, 0f, 0f), new Vector3(0f, 0f, 0f), new Vector3(0f, 0f, 0f), new Vector3(0f, 0f, 0f), time)
     {
     }
